requirements.txt
requests
openai
langchain
python-dotenv
streamlit